requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
confuse==2.0.1
cryptography==41.0.5
distlib==0.3.7
Django==4.2.7
django-crispy-forms==2.1
django-paypal==2.0
django-tinymce==3.6.1
filelock==3.12.4
fontawesomefree==6.4.2
idna==3.4
Pillow==10.1.0
platformdirs==3.10.0
pycparser==2.21
python-paypal-api==0.1.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.5